+++ tests/ClientTest.php
@@ -1261,13 +1261,13 @@ class ClientTest extends TestCase
         $this->client->addFeatureFlag('another name', 'with variant');
 
         $expected = [
-            ['featureFlag' => 'a name'],
-            ['featureFlag' => 'another name', 'variant' => 'with variant'],
+            new FeatureFlag('a name'),
+            new FeatureFlag('another name', 'with variant'),
         ];
 
         $actual = $this->client->getConfig()->getFeatureFlagsCopy()->toArray();
 
-        $this->assertSame($expected, $actual);
+        $this->assertEquals($expected, $actual);
     }
 
     public function testMultipleFeatureFlagsCanBeAddedToClientAtOnce()
@@ -1280,15 +1280,15 @@ class ClientTest extends TestCase
         ]);
 
         $expected = [
-            ['featureFlag' => 'a name'],
-            ['featureFlag' => 'another name', 'variant' => 'with variant'],
-            ['featureFlag' => 'name3'],
-            ['featureFlag' => 'four', 'variant' => 'yes'],
+            new FeatureFlag('a name'),
+            new FeatureFlag('another name', 'with variant'),
+            new FeatureFlag('name3'),
+            new FeatureFlag('four', 'yes'),
         ];
 
         $actual = $this->client->getConfig()->getFeatureFlagsCopy()->toArray();
 
-        $this->assertSame($expected, $actual);
+        $this->assertEquals($expected, $actual);
     }
 
     public function testAFeatureFlagCanBeRemovedFromClient()
@@ -1299,12 +1299,12 @@ class ClientTest extends TestCase
         $this->client->clearFeatureFlag('another name');
 
         $expected = [
-            ['featureFlag' => 'a name'],
+            new FeatureFlag('a name'),
         ];
 
         $actual = $this->client->getConfig()->getFeatureFlagsCopy()->toArray();
 
-        $this->assertSame($expected, $actual);
+        $this->assertEquals($expected, $actual);
     }
 
     public function testAllFeatureFlagsCanBeRemovedFromClient()
@@ -1342,6 +1342,61 @@ class ClientTest extends TestCase
                 $payload = $this->getPayloadFromGuzzleOptions($options);
 
                 $this->assertTrue(isset($payload['events'][0]['featureFlags']));
+
+                $expected = [
+                    ['featureFlag' => 'added from client'],
+                    ['featureFlag' => 'added from report 1', 'variant' => 'yes'],
+                    ['featureFlag' => 'added from report 2'],
+                ];
+
+                $this->assertSame($expected, $payload['events'][0]['featureFlags']);
+
+                return true;
+            }
+        );
+
+        $this->client->flush();
+    }
+
+    public function testFeatureFlagsCanBeAccessedInCallbacks()
+    {
+        $this->client->registerCallback(function (Report $report) {
+            $report->addFeatureFlag('hi');
+
+            $report->addFeatureFlags([
+                new FeatureFlag('added from report 1', 'yes'),
+                new FeatureFlag('added from report 2'),
+            ]);
+
+            $report->clearFeatureFlag('hello');
+            $report->clearFeatureFlag('hi');
+        });
+
+        $called = false;
+        $this->client->registerCallback(function (Report $report) use (&$called) {
+            $expected = [
+                new FeatureFlag('added from client'),
+                new FeatureFlag('added from report 1', 'yes'),
+                new FeatureFlag('added from report 2'),
+            ];
+
+            $this->assertEquals($expected, $report->getFeatureFlags());
+
+            $called = true;
+        });
+
+        $this->client->addFeatureFlag('added from client');
+        $this->client->addFeatureFlag('hello');
+
+        $this->client->notifyError('SomeError', 'Some message');
+
+        $this->expectGuzzlePostWithCallback(
+            $this->client->getNotifyEndpoint(),
+            function ($options) use ($called) {
+                $payload = $this->getPayloadFromGuzzleOptions($options);
+
+                $this->assertTrue(isset($payload['events'][0]['featureFlags']));
+                $this->assertTrue($called);
 
                 $expected = [
                     ['featureFlag' => 'added from client'],